[Excel Analyzer] Extract metadata parsing
Move metadata parsing into its own class, this saves passing around the
ZIP file to individual methods.
